Placed methods in respective classes. changed user to employee. made crud on project and subproject.
