+++ src/main/java/com/alphaS/alphasolutions/controllers/TaskCompassController.java
@@ -3,10 +3,11 @@ package com.alphaS.alphasolutions.controllers;
 
 
 import com.alphaS.alphasolutions.model.ProjectModel;
-import com.alphaS.alphasolutions.model.SubProjectModel;
-import com.alphaS.alphasolutions.model.UserModel;
+import com.alphaS.alphasolutions.model.EmployeeModel;
+import com.alphaS.alphasolutions.repositories.ClientRepository;
 import com.alphaS.alphasolutions.repositories.ProjectRepository;
-import com.alphaS.alphasolutions.repositories.UserRepository;
+import com.alphaS.alphasolutions.repositories.TeamRepository;
+import com.alphaS.alphasolutions.repositories.EmployeeRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.http.HttpStatus;
@@ -22,12 +23,17 @@ import java.util.List;
 @Controller
 public class TaskCompassController {
 
-    private final UserRepository userRepository;
+    private final EmployeeRepository employeeRepository;
     private final ProjectRepository projectRepository;
+    private final ClientRepository clientRepository;
+    private final TeamRepository teamRepository;
 
-    public TaskCompassController(UserRepository userRepository, ProjectRepository projectRepository) {
-        this.userRepository = userRepository;
+
+    public TaskCompassController(EmployeeRepository employeeRepository, ProjectRepository projectRepository, ClientRepository clientRepository, TeamRepository teamRepository) {
+        this.employeeRepository = employeeRepository;
         this.projectRepository = projectRepository;
+        this.clientRepository = clientRepository;
+        this.teamRepository = teamRepository;
     }
 
     @GetMapping("")
@@ -42,8 +48,8 @@ public class TaskCompassController {
     @PostMapping("/signin")
     public String signinpostmapping(@RequestParam("username") String username, @RequestParam("password") String password, HttpSession session, Model model){
         try {
-            UserModel usermodel = userRepository.signin(username, password);
-            session.setAttribute("user", usermodel); // store user in session
+            EmployeeModel employeeModel = employeeRepository.signin(username, password);
+            session.setAttribute("user", employeeModel); // store user in session
             return "dashboard"; // return dashboard page if login is successful
         } catch (SQLException e) {
             model.addAttribute("error", "Username or password incorrect"); // add error message to model
@@ -67,31 +73,23 @@ public class TaskCompassController {
                                                @RequestParam String country,
                                                @RequestParam String clientId) {
         try {
-            String result = projectRepository.createClient(clientName, contactPoNo, contactPerson, companyPoNo, address, zipCode, country, clientId);
+            String result = clientRepository.createClient(clientName, contactPoNo, contactPerson, companyPoNo, address, zipCode, country, clientId);
             return ResponseEntity.ok(result);
         } catch (SQLException e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to create client");
         }
     }
 
-    @GetMapping("/createproject")
-    public String CreateProject(Model model) {
-        model.addAttribute("project", new ProjectModel());
-        model.addAttribute("subProject", new SubProjectModel());
-        return "createproject";
-    }
-
-    @PostMapping("/createproject") //TODO: SESSIONS
-    public ResponseEntity<String> createProject(@RequestBody ProjectModel project, @RequestBody(required = false) SubProjectModel subProject) {
+    @GetMapping("/clients/search")
+    @ResponseBody
+    public ResponseEntity<String> searchClient(@RequestParam String clientName) {
         try {
-            projectRepository.createProject(project.getProjectName(), project.getProjectDescription(), project.getStartDate(), project.getEndDate(), subProject);
-            return ResponseEntity.ok("Project successfully created");
+            String result = clientRepository.searchClients(clientName).toString();
+            return ResponseEntity.ok(result);
         } catch (SQLException e) {
-            e.printStackTrace();
-            return ResponseEntity.badRequest().body("Failed to create project");
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to search client");
         }
     }
-
 
     @PostMapping("/clients/edit")
     public ResponseEntity<String> editClient(@RequestParam String clientName,
@@ -103,31 +101,34 @@ public class TaskCompassController {
                                              @RequestParam String country,
                                              @RequestParam String clientId) {
         try {
-            String result = projectRepository.editClient(clientName, contactPoNo, contactPerson, companyPoNo, address, zipCode, country, clientId);
-            return ResponseEntity.ok(result);
-            } catch (SQLException e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to edit client");
-        }
-
-    }
-
-    @GetMapping("/clients/search")
-    @ResponseBody
-    public ResponseEntity<String> searchClient(@RequestParam String clientName) {
-        try {
-            String result = projectRepository.searchClients(clientName).toString();
+            String result = clientRepository.editClient(clientName, contactPoNo, contactPerson, companyPoNo, address, zipCode, country, clientId);
             return ResponseEntity.ok(result);
         } catch (SQLException e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to search client");
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to edit client");
         }
     }
-
     @PostMapping("/clients/delete")
     public String deleteClient(@RequestParam int clientID){
-        String message = projectRepository.deleteClient(clientID);
+        String message = clientRepository.deleteClient(clientID);
         return message;
     }
 
+    @GetMapping("/createproject")
+    public String CreateProject(Model model) {
+        model.addAttribute("project", new ProjectModel());
+        return "createproject";
+    }
+
+    @PostMapping("/createproject") //TODO: SESSIONS
+    public ResponseEntity<String> createProject(@RequestBody ProjectModel project) {
+        try {
+            projectRepository.createProject(project.getProjectName(), project.getProjectDescription(), project.getStartDate(), project.getEndDate());
+            return ResponseEntity.ok("Project successfully created");
+        } catch (SQLException e) {
+            e.printStackTrace();
+            return ResponseEntity.badRequest().body("Failed to create project");
+        }
+    }
 
     @GetMapping("/project/search")
     @ResponseBody
@@ -150,21 +151,21 @@ public class TaskCompassController {
     public ResponseEntity<String> createTeam(HttpServletRequest request) {
         try {
             String teamName = request.getParameter("teamName");
-            int teamId = projectRepository.createTeam(teamName);
+            int teamId = teamRepository.createTeam(teamName);
             return ResponseEntity.ok().body(Integer.toString(teamId));
         } catch (SQLException e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to create team");
         }
     }
 
-    @PostMapping("/addTeamMembers")
+    @PostMapping("/addEmployeeToTeam")
     @ResponseBody
-    public ResponseEntity<String> addTeamMembers(HttpServletRequest request) {
+    public ResponseEntity<String> addEmployeeToTeam(HttpServletRequest request) {
         try {
             int teamId = Integer.parseInt(request.getParameter("teamId"));
             String[] usernames = request.getParameterValues("usernames[]");
             List<String> userNames = Arrays.asList(usernames);
-            projectRepository.addTeamMembers(teamId, userNames);
+            teamRepository.AddEmployeeToTeam(teamId, userNames);
             return ResponseEntity.ok().body("Team members added successfully.");
         } catch (SQLException e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to add team members.");
