+++ src/main/java/com/alphaS/alphasolutions/model/SubProjectModel.java
@@ -10,7 +10,6 @@ public class SubProjectModel {
     }
 
     public SubProjectModel() {
-
     }
 
     public String getSubProjectName() {
