+++ src/main/java/com/alphaS/alphasolutions/repositories/TaskRepository.java
@@ -1,4 +1,9 @@
 package com.alphaS.alphasolutions.repositories;
 
+import org.springframework.stereotype.Repository;
+
+@Repository
 public class TaskRepository {
+
+
 }
